extracted ScriptHelper
